+++ R/pip_gdp_update.R
@@ -5,10 +5,12 @@
 #' @inheritParams pip_gdp
 #' @inheritParams pip_wdi_update
 #' @keywords internal
-pip_gdp_update <- function(force = FALSE,
-                           maindir = gls$PIP_DATA_DIR,
-                           sna_tag = "main",
-                           from    = "file") {
+pip_gdp_update <- function(force      = FALSE,
+                           maindir    = gls$PIP_DATA_DIR,
+                           sna_branch = c("main", "dev"),
+                           from       = "file") {
+
+  sna_branch <- match.arg(sna_branch)
 
   #----------------------------------------------------------
   #   Load data
@@ -28,8 +30,8 @@ pip_gdp_update <- function(force = FALSE,
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   ## Special national accounts --------
-  usna <- glue("https://github.com/PIP-Technical-Team/pip-sna/raw/{sna_tag}/sna.csv")
-  umet <- "https://github.com/PIP-Technical-Team/pip-sna/raw/main/sna_metadata.csv"
+  usna <- glue("https://github.com/PIP-Technical-Team/pip-sna/raw/{sna_branch}/sna.csv")
+  umet <- glue("https://github.com/PIP-Technical-Team/pip-sna/raw/{sna_branch}/sna_metadata.csv")
 
   tryCatch(
     expr = {
@@ -45,9 +47,9 @@ pip_gdp_update <- function(force = FALSE,
       tags  <- c("main", get_gh_tags(owner, repo))
 
 
-      if (! (sna_tag  %in% tags)) {
+      if (! (sna_branch  %in% tags)) {
         msg     <- c(
-          "{.field sna_tag} specified ({sna_tag}) does not exist in repo
+          "{.field sna_branch} specified ({sna_branch}) does not exist in repo
           {.file {owner}/{repo}}",
           "i" = "Select one among {.field {tags}}"
           )