+++ NEWS.md
@@ -1,3 +1,5 @@
+# pipaux (development version)
+
 # pipaux 0.0.7
 
 -   Add `ppp_vintage` database when `ppp` is updated.