+++ NEWS.md
@@ -1,4 +1,4 @@
-# pipaux (development version)
+# pipaux 0.0.8
 
 -   Retrieve SNA data from github repo, not from drive anymore